requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
airfold-cli==0.14.3
airfold-common==0.14.3
annotated-types==0.7.0
attrs==25.3.0
boto3==1.38.18
botocore==1.38.18
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
deepdiff==8.5.0
duckdb==1.2.2
Faker==37.3.0
frozenlist==1.6.0
humanize==4.12.3
idna==3.10
jmespath==1.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
merge-args==0.1.5
multidict==6.4.3
orderly-set==5.4.1
propcache==0.3.1
pyarrow==20.0.0
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
readchar==4.2.1
requests==2.32.3
rich==14.0.0
s3transfer==0.12.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
sqlglot==26.17.1
tenacity==9.1.2
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
unidiff==0.7.5
urllib3==2.4.0
yarl==1.20.0
